feat: refactoring and datatypes visualization WIP
