+++ app/datasources/page.tsx
@@ -7,7 +7,7 @@ export default async function Home() {
 
   return (
     <main className="h-lvh flex">
-      <SidePanel />
+      <SidePanel sidePanelType="datasources" />
       <div id="frames-area" className="flex-1 flex flex-col gap-2 p-2 h-full">
         <HeaderFrame />
         <FramesContent />
